+++ src/odin/__init__.py
@@ -1,0 +1,2 @@
+"""ODIN Hermes Alive package."""
+