+++ src/odin/hermes/__init__.py
@@ -1,0 +1,2 @@
+"""Hermes package."""
+